requirements.txt
streamlit==1.41.1
langchain-groq==0.2.3
chromadb==0.5.23
pandas==2.0.3
langchain-community==0.2.19
langchain-core==2.3.43
python-dotenv==1.0.1
